Conflicts:
	cheatingday/src/main/java/com/icia/cheatingday/review/dao/ReviewDao.java
	cheatingday/src/main/resources/mapper/reviewMapper.xml
	cheatingday/src/main/resources/mapper/storeMapper.xml
	cheatingday/target/classes/com/icia/cheatingday/manager/service/MReviewOrderService.class
	cheatingday/target/classes/com/icia/cheatingday/review/dao/ReviewDao.class
	cheatingday/target/classes/mapper/reviewMapper.xml
	cheatingday/target/classes/mapper/storeMapper.xml
	cheatingday/target/m2e-wtp/web-resources/META-INF/maven/com.icia/cheatingday/pom.properties

+++ cheatingday/src/main/java/com/icia/cheatingday/admin/service/rest/AdminRestService.java
@@ -4,9 +4,8 @@ import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
 
 import com.icia.cheatingday.admin.dao.*;
-import com.icia.cheatingday.manager.entity.*;
+import com.icia.cheatingday.exception.*;
 import com.icia.cheatingday.review.dao.*;
-import com.icia.cheatingday.user.dao.*;
 import com.icia.cheatingday.user.entity.*;
 
 @Service
@@ -19,6 +18,8 @@ public class AdminRestService {
 
 	public void deleteReport(int rNo) {
 		Review review = rdao.findById(rNo);
+		if(review==null)
+			throw new JobFailException("해당 리뷰를 찾을 수 없습니다");
 		rdao.delete(rNo);
 		
 	}
