+++ cheatingday/src/main/java/com/icia/cheatingday/center/service/rest/QnARestService.java
@@ -4,15 +4,19 @@ import java.io.*;
 import java.time.*;
 import java.time.format.*;
 import java.util.*;
+import java.util.regex.*;
+
 import org.modelmapper.*;
 import org.springframework.beans.factory.annotation.*;
 import org.springframework.stereotype.*;
+import org.springframework.util.*;
 import org.springframework.web.multipart.*;
+
 import com.fasterxml.jackson.databind.*;
-import com.icia.cheatingday.admin.dao.*;
 import com.icia.cheatingday.center.dao.*;
 import com.icia.cheatingday.center.dto.*;
 import com.icia.cheatingday.center.entity.*;
+import com.icia.cheatingday.exception.*;
 import com.icia.cheatingday.manager.dao.*;
 
 @Service
@@ -26,16 +30,20 @@ public class QnARestService {
 	@Autowired
 	private ManagerDao mdao;
 	@Autowired
-	private AdminDao adao;
-	@Autowired
 	private ModelMapper mapper;
-	@Value("http://localhost:8081/ckimage/")
-	private String ckUrl;
+	@Value("${imageFolder}")
+	private String imageFolder;
+	@Value("${imagePath}")
+	private String imagePath;
 	@Autowired
 	private ObjectMapper objectmapper;
 	
+	Pattern ckImagePattern = Pattern.compile("src=\".+\"\\s");
+	
 	public QnADto.DtoForRead read(Integer qNo, String username){
 		QnA qna = qdao.findById(qNo);
+		if(qna==null)
+			throw new JobFailException("해당 문의를 찾을수 없습니다");
 		QnADto.DtoForRead dto = mapper.map(qna,QnADto.DtoForRead.class);
 		String str = qna.getQWriteTime().format(DateTimeFormatter.ofPattern("yyyy년 MM월 dd일"));
 		dto.setMUsername(mdao.findMusernameByMnum(dto.getMNum()));
@@ -56,6 +64,8 @@ public class QnARestService {
 	}
 	public List<QnAComment> deleteComment(Integer qNo, Integer qcNo, String username) {
 		QnAComment qnAComment = qndao.findById(qcNo);
+		if(qnAComment.getAUsername().equals(username)==false)
+			throw new JobFailException("댓글을 찾을 수 없습니다");
 		qndao.delete(qcNo);
 		qdao.update(QnA.builder().qNo(qnAComment.getQNo()).qIscomment(false).build());
 		return qndao.findAllByQno(qNo);
@@ -63,7 +73,11 @@ public class QnARestService {
 	
 	public void deletQna(Integer qNo, String username) {
 		QnA qnA = qdao.findById(qNo);
-		/*String content = qnA.getQContent();
+		if(qnA==null)
+			throw new JobFailException("헤당 문의를 찾을 수 없습니다");	
+		if(mdao.findMusernameByMnum(qnA.getMNum()).equals(username)==false)
+			throw new UserNotFoundException();
+		String content = qnA.getQContent();
 		Matcher matcher = ckImagePattern.matcher(content);
 		while(matcher.find()) {
 			String src = matcher.group();
@@ -73,11 +87,16 @@ public class QnARestService {
 			File file = new File(imageFolder, fileName);
 			if(file.exists()==true)
 				file.delete();
-		}*/
+		}
 		qdao.delete(qNo);
 	}
 	public void updateQnA(QnADto.DtoForUpdate dto) {
 		QnA qna = qdao.findById(dto.getQNo());
+		if(qna==null)
+			throw new JobFailException("문의를 찾을 수 없습니다");
+		dto.setMUsername(mdao.findMirumeByMnum(qna.getMNum()));
+		if(mdao.findMirumeByMnum(qna.getMNum()).equals(dto.getMUsername())==false)
+			throw new UserNotFoundException();
 		qna = mapper.map(dto, QnA.class);
 		qdao.update(qna);
 	}
@@ -85,33 +104,24 @@ public class QnARestService {
 	public void updateQnAcomment(QnAComment qnAComment){
 		 qndao.update(qnAComment);
 	}
-	public String saveCkImage(MultipartFile upload) {
-	      Map<String, String> map = new HashMap<String, String>();
-	      if (upload != null) {
-	    	  System.out.println(upload);
-	    	  System.out.println(upload);
-	    	  System.out.println(upload);
-	         if (upload.getContentType().toLowerCase().startsWith("image/")) {
-	            String imageName = UUID.randomUUID().toString() + ".jpg";
-	            System.out.println(imageName);
-	            System.out.println(imageName);
-	            System.out.println(imageName);
-	            try {
-	               File file = new File("d:/upload/ckimage", imageName);
-	               // FileCopyUtils.copy(upload.getBytes(), file);
-	               upload.transferTo(file);
-	               map.put("uploaded", "1");
-	               map.put("fileName", imageName);
-	               map.put("url", ckUrl + imageName);
-	               return objectmapper
-	                     .writerWithDefaultPrettyPrinter() // 줄바꿈
-	                     .writeValueAsString(map);
-	            } catch (Exception e) {
-	               e.printStackTrace();
-	            }
-	         }
-	      }
-	      return null;
-	   }
+	public String saveCkImage(MultipartFile upload) throws IOException {
+		Map<String,String> map = new HashMap<String, String>();
+		if(upload!=null) {
+			if(upload.getContentType().toLowerCase().startsWith("image/")){
+				String imageName = UUID.randomUUID().toString() + ".jpg";
+				File file = new File(imageFolder, imageName);
+				FileCopyUtils.copy(upload.getBytes(), file);
+				String fileUrl = imagePath + imageName;
+				// json 데이터로 등록
+				// {"uploaded" : 1, "fileName" : "test.jpg", "url" : "/img/test.jpg"}
+				// 이런 형태로 리턴이 나가야함.
+				map.put("uploaded", "1");
+				map.put("fileName", imageName);
+				map.put("url", fileUrl);
+                return objectmapper.writerWithDefaultPrettyPrinter().writeValueAsString(map);
+			}
+		}
+		return null;
+	}
 	
 }
